refactor: move ibc-events into the relevant voyager cosmos module
the library now only exports the macro to define event types
